src/main.py
#!/usr/bin/env python3
"""
Autonomous Office Protocol (AOP) - Main Agent

Colosseum Agent Hackathon submission demonstrating:
- Autonomous AI decision-making with Claude
- On-chain activity logging via AgentWallet
- DeFi position monitoring on Solana
- Multi-agent coordination
"""

import os
import asyncio
import time
from datetime import datetime
from dotenv import load_dotenv
from anthropic import Anthropic

# Load environment variables
load_dotenv()

# Import local modules
from agentwallet import AgentWallet

class AutonomousOfficeAgent:
    """Main autonomous agent for Colosseum hackathon"""
    
    def __init__(self):
        """Initialize agent with Claude AI and AgentWallet"""
        self.anthropic_api_key = os.getenv("ANTHROPIC_API_KEY")
        self.colosseum_api_key = os.getenv("COLOSSEUM_API_KEY")
        
        if not self.anthropic_api_key:
            raise ValueError("ANTHROPIC_API_KEY not set in environment")
        
        if not self.colosseum_api_key:
            print("⚠️  COLOSSEUM_API_KEY not set - on-chain logging disabled")
        
        # Initialize components
        self.claude = Anthropic(api_key=self.anthropic_api_key)
        self.wallet = AgentWallet(api_key=self.colosseum_api_key)
        
        # Agent state
        self.running = False
        self.activity_count = 0
        
        print("🤖 Autonomous Office Protocol initialized")
        print(f"📊 Claude AI: Connected")
        print(f"🔗 AgentWallet: {'Connected' if self.colosseum_api_key else 'Offline'}")
    
    async def monitor_positions(self):
        """Monitor DeFi positions on Solana"""
        print("\n👀 Monitoring Solana DeFi positions...")
        
        # Simulate position monitoring
        # In production, this would query Solend, Kamino, Marinade
        positions = {
            "solend": {
                "collateral_usd": 10000,
                "debt_usd": 6000,
                "health_factor": 1.45
            },
            "kamino": {
                "collateral_usd": 5000,
                "debt_usd": 2500,
                "health_factor": 1.85
            }
        }
        
        # Log monitoring activity
        self.wallet.log_activity_onchain(
            activity_type="position_monitor",
            data={
                "timestamp": datetime.utcnow().isoformat(),
                "protocols_checked": ["solend", "kamino", "marinade"],
                "positions_found": len(positions),
                "status": "healthy"
            }
        )
        
        self.activity_count += 1
        print(f"✅ Positions monitored (Activity #{self.activity_count})")
        
        return positions
    
    async def analyze_risk_with_claude(self, positions):
        """Use Claude AI to analyze risk and recommend actions"""
        print("\n🧠 Analyzing risk with Claude AI...")
        
        # Prepare prompt for Claude
        prompt = f"""You are an AI agent analyzing DeFi positions on Solana for liquidation risk.

Current positions:
{positions}

Analyze the health factors and provide:
1. Risk assessment (low/medium/high)
2. Recommended action (monitor/rebalance/urgent)
3. Reasoning for your decision

Be concise and actionable."""
        
        try:
            # Call Claude API
            message = self.claude.messages.create(
                model="claude-3-5-sonnet-20241022",
                max_tokens=1024,
                messages=[{
                    "role": "user",
                    "content": prompt
                }]
            )
            
            analysis = message.content[0].text
            
            # Log AI analysis activity
            self.wallet.log_activity_onchain(
                activity_type="ai_analysis",
                data={
                    "timestamp": datetime.utcnow().isoformat(),
                    "model": "claude-3-5-sonnet-20241022",
                    "positions_analyzed": len(positions),
                    "analysis": analysis[:200],  # First 200 chars
                    "status": "complete"
                }
            )
            
            self.activity_count += 1
            print(f"✅ Risk analyzed by Claude (Activity #{self.activity_count})")
            print(f"\n📋 Analysis:\n{analysis}\n")
            
            return analysis
            
        except Exception as e:
            print(f"❌ Claude analysis failed: {e}")
            return None
    
    async def execute_action(self, action_type):
        """Execute recommended action"""
        print(f"\n⚡ Executing action: {action_type}")
        
        # Simulate action execution
        # In production, this would interact with Solana programs
        
        actions = {
            "monitor": "Continue monitoring positions",
            "rebalance": "Execute flash loan rebalancing",
            "urgent": "Emergency position closure"
        }
        
        result = actions.get(action_type, "Unknown action")
        
        # Log execution activity
        self.wallet.log_activity_onchain(
            activity_type="action_execution",
            data={
                "timestamp": datetime.utcnow().isoformat(),
                "action_type": action_type,
                "result": result,
                "status": "success"
            }
        )
        
        self.activity_count += 1
        print(f"✅ Action executed (Activity #{self.activity_count})")
        
        return result
    
    async def run_cycle(self):
        """Run one complete agent cycle"""
        print("\n" + "=" * 60)
        print(f"🔄 AGENT CYCLE - {datetime.utcnow().strftime('%Y-%m-%d %H:%M:%S UTC')}")
        print("=" * 60)
        
        # Step 1: Monitor positions
        positions = await self.monitor_positions()
        
        # Step 2: Analyze with Claude
        analysis = await self.analyze_risk_with_claude(positions)
        
        # Step 3: Execute action
        if analysis:
            # Simple decision logic based on analysis
            if "urgent" in analysis.lower():
                await self.execute_action("urgent")
            elif "rebalance" in analysis.lower():
                await self.execute_action("rebalance")
            else:
                await self.execute_action("monitor")
        
        print("\n" + "=" * 60)
        print(f"✅ Cycle complete - Total activities: {self.activity_count}")
        print("=" * 60)
    
    async def run(self, interval=600):
        """Run agent continuously"""
        self.running = True
        
        print("\n🚀 Starting Autonomous Office Protocol agent...")
        print(f"⏰ Cycle interval: {interval} seconds ({interval/60} minutes)")
        print(f"🎯 Target: 500+ activities over 3 days")
        print(f"📊 Required rate: ~7 activities/hour")
        print("\nPress Ctrl+C to stop\n")
        
        try:
            while self.running:
                await self.run_cycle()
                
                # Wait for next cycle
                print(f"\n⏳ Waiting {interval} seconds until next cycle...")
                await asyncio.sleep(interval)
                
        except KeyboardInterrupt:
            print("\n\n🛑 Agent stopped by user")
            self.running = False
        
        print(f"\n📊 Final Statistics:")
        print(f"   Total activities logged: {self.activity_count}")
        print(f"   Runtime: {datetime.utcnow().strftime('%Y-%m-%d %H:%M:%S UTC')}")

async def main():
    """Main entry point"""
    print("=" * 60)
    print("🏆 AUTONOMOUS OFFICE PROTOCOL")
    print("   Colosseum Agent Hackathon 2026")
    print("=" * 60)
    print()
    
    # Initialize and run agent
    agent = AutonomousOfficeAgent()
    
    # Run with 10-minute cycles (600 seconds)
    # This generates ~6 activities per hour (2 per cycle)
    await agent.run(interval=600)

if __name__ == "__main__":
    asyncio.run(main())
