requirements.txt
# Core Dependencies
anthropic>=0.18.0
requests>=2.31.0
python-dotenv>=1.0.0

# Cryptography
cryptography>=42.0.0

# Solana
solana>=0.30.0
anchorpy>=0.18.0

# Web3
web3>=6.15.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Development
black>=24.1.0
flake8>=7.0.0
mypy>=1.8.0
